+++ Course.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Wild_Project1
+{
+    internal class Course
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}
