+++ Grades.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Wild_Project1
+{
+    internal class Grades
+    {
+        public int CourseId { get; set; }
+        public double Value { get; set; }
+        public string Commentary { get; set; }
+    }
+}
